Refactored getUsers into MVC
