+++ src/models/models.ts
@@ -1,0 +1,6 @@
+import model from "../schemas/model";
+
+export const selectUsers = async () => {
+    const users = await model.find()
+    return users 
+}
